// src/interface/connect_us.interface.ts
export interface IConnectUs {
    phone_1: string;
    phone_2?: string | null;
    phone_3?: string | null;
    phone_4?: string | null;
    email: string;
    address: string;
    map_embed_link?: string;
    facebook_link?: string;
    instagram_link?: string;
    whatsapp_link?: string;
    youtube_link?: string;
    tiktok_link?: string;
}